CWA detection (experimental)
