+++ platformio.ini
@@ -37,16 +37,16 @@ halfile = generic.h
 
 [platformio]
 ; upload firmware to board with usb cable
-;default_envs = usb
+default_envs = usb
 ; upload firmware to a jfrog bintray repository
 ;default_envs = ota
 ; use latest versions of libraries
-default_envs = dev
+;default_envs = dev
 description = Paxcounter is a device for metering passenger flows in realtime. It counts how many mobile devices are around.
 
 [common]
 ; for release_version use max. 10 chars total, use any decimal format like "a.b.c"
-release_version = 1.9.996
+release_version = 2.0.1
 ; DEBUG LEVEL: For production run set to 0, otherwise device will leak RAM while running!
 ; 0=None, 1=Error, 2=Warn, 3=Info, 4=Debug, 5=Verbose
 debug_level = 4
@@ -61,19 +61,19 @@ display_library = ; set by build.py and taken from hal file
 lib_deps_lora =
     MCCI LoRaWAN LMIC library@3.2.0 ; MCCI LMIC by Terrill Moore
 lib_deps_display =
-    OneBitDisplay@1.5.0
+    OneBitDisplay@1.7.1
     QRCode@0.0.1
     BitBang_I2C@2.1.1
-    TFT_eSPI@>=2.2.8
+    TFT_eSPI@>=2.2.18
 lib_deps_ledmatrix =
     Ultrathin_LED_Matrix@>=1.0.0
 lib_deps_rgbled =
-    SmartLeds@>=1.2.0
+    SmartLeds@>=1.2.1
 lib_deps_gps =
     1655@>=1.0.2 ; #1655 TinyGPSPlus by Mikal Hart
 lib_deps_sensors =
     Adafruit Unified Sensor@>=1.1.4
-    Adafruit BME280 Library@>=2.0.2
+    Adafruit BME280 Library@>=2.1.0
     Adafruit BMP085 Library@>=1.1.0
     BSEC Software Library@1.5.1474
     https://github.com/ricki-z/SDS011.git
