+++ src/display.cpp
@@ -265,11 +265,10 @@ void dp_drawPage(time_t t, bool nextpage) {
     else
       dp_printf("WIFI:off");
     if (cfg.blescan)
-#if !(COUNT_CWA)
+#if (!COUNT_ENS)
       dp_printf("BLTH:%-5d", macs_ble);
 #else
-      dp_printf("BLTH:%-5d", macs_ble);
-      dp_printf("(CWA:%d)", cwa_report());
+      dp_printf(" CWA:%-5d", cwa_report());
 #endif
     else
       dp_printf(" BLTH:off");
@@ -281,10 +280,10 @@ void dp_drawPage(time_t t, bool nextpage) {
 #elif ((!WIFICOUNTER) && (BLECOUNTER))
     if (cfg.blescan) {
       dp_printf("BLTH:%-5d", macs_ble);
-#if (COUNT_CWA)
+#if (COUNT_ENS)
       dp_printf("(CWA:%d)", cwa_report());
-#endif      
-  } else
+#endif
+    } else
       dp_printf("BLTH:off");
 #else
     dp_printf("Sniffer disabled");
@@ -357,7 +356,7 @@ void dp_drawPage(time_t t, bool nextpage) {
 
 #if (HAS_LORA)
 
-    // 3|NtwkID:000000 TXpw:aa
+    // 3|NetID:000000 PWR:aa
     // 4|DevAdd:00000000  DR:0
     // 5|CHMsk:0000 Nonce:0000
     // 6|CUp:000000 CDn:000000
@@ -365,8 +364,7 @@ void dp_drawPage(time_t t, bool nextpage) {
 
     dp_setFont(MY_FONT_SMALL);
     dp_setTextCursor(0, 3);
-    dp_printf("NetwID:%06X TXpw:%-2d", LMIC.netid & 0x001FFFFF,
-              LMIC.radio_txpow);
+    dp_printf("NetID:%06X PWR:%-2d", LMIC.netid & 0x001FFFFF, LMIC.radio_txpow);
     dp_println();
     dp_printf("DevAdd:%08X DR:%1d", LMIC.devaddr, LMIC.datarate);
     dp_println();
