+++ src/paxcounter.conf
@@ -16,7 +16,7 @@
 
 // Set this to include BLE counting and vendor filter functions, or to switch off WIFI counting
 #define VENDORFILTER                    1       // set to 0 if you want to count things, not people
-#define BLECOUNTER                      0       // set it to 1 if you want to use BLE count, at expense of power & memory
+#define BLECOUNTER                      1       // set it to 1 if you want to use BLE count, at expense of power & memory
 #define WIFICOUNTER                     1       // set it to 0 if you want to switch off WIFI count
 
 // BLE scan parameters
@@ -24,11 +24,12 @@
 #define BLESCANWINDOW                   80      // [milliseconds] scan window, see below, 3 .. 10240, default 80ms
 #define BLESCANINTERVAL                 80      // [illiseconds] scan interval, see below, 3 .. 10240, default 80ms = 100% duty cycle
 
-#define COUNT_CWA                       0       // count found copies of the Corona Warn App (CWA)
-                                                // set to 0 if you do not want to count these apps
+// Corona Exposure Notification Service(ENS) counter
+#define COUNT_ENS                       1       // count found number of devices which advertise Exposure Notification Service
+                                                // set to 0 if you do not want to count these devices
 
 // for additional sensors (added by some user)
-#define HAS_SENSOR_1                    0       // set to 1 if you want to count CWAs
+#define HAS_SENSOR_1                    1       // set to 1 if you want to count CWAs
 #define HAS_SENSOR_2                    0       // not used 
 #define HAS_SENSOR_3                    0       // not used
 #define HAS_SENSORS     (HAS_SENSOR_1 || HAS_SENSOR_2 || HAS_SENSOR_3)   // to simplify things
@@ -82,8 +83,8 @@
 #define RESPONSE_TIMEOUT_MS             60000   // firmware binary server connection timeout [milliseconds]
 
 // settings for syncing time of node with a time source (network / gps / rtc / timeserver)
-#define TIME_SYNC_LORAWAN               1       // set to 1 to use LORA network as time source, 0 means off [default = 1]
-#define TIME_SYNC_LORASERVER            0       // set to 1 to use LORA timeserver as time source, 0 means off [default = 0]
+#define TIME_SYNC_LORAWAN               0       // set to 1 to use LORA network as time source, 0 means off [default = 1]
+#define TIME_SYNC_LORASERVER            1       // set to 1 to use LORA timeserver as time source, 0 means off [default = 0]
 #define TIME_SYNC_INTERVAL              60      // sync time attempt each .. minutes from time source [default = 60], 0 means off
 #define TIME_SYNC_INTERVAL_RETRY        10       // retry time sync after lost sync each .. minutes [default = 10], 0 means off
 #define TIME_SYNC_SAMPLES               1       // number of time requests for averaging, max. 255
